requirements.txt
requests
urllib3
pyzabbix
pytest
pytest-lazy-fixtures
python-dotenv
discord.py
flask

